Fixes #21.

This turned out to be far simpler than the cache in Splint. Converting the command line arguments into a context is a pure function (albeit one that can fail). So I only needed an `IORef` rather than a `TVar`.