@tfausak 2024-04-17
Curiously I haven't been able to prove that this is actually any faster. I wrote a little script to generate 1,000 modules that use `Map.*` identifiers:

``` hs
main = do
  let count = 1000
  mapM_
    ( \n ->
        writeFile ("library/M" <> show n <> ".hs") $
          unlines
            [ "module M" <> show n <> " where",
              "x :: Map.Map Char Int",
              "x = Map.empty"
            ]
    )
    [1 .. count]
  writeFile "library/A.hs"
    . unlines
    $ "module A where"
      : fmap
        (\n -> "import M" <> show n <> " ()")
        [1 .. count]
```

Then I build that as a package with a bunch of aliases set up:

``` cabal
  ghc-options:
    -fplugin=Imp
    -fplugin-opt=Imp:--alias=_:This
    -fplugin-opt=Imp:--alias=Control.Applicative:Applicative
    -fplugin-opt=Imp:--alias=Control.Concurrent:Concurrent
    -fplugin-opt=Imp:--alias=Control.Exception:Exception
    -fplugin-opt=Imp:--alias=Control.Monad:Monad
    -fplugin-opt=Imp:--alias=Data.Bool:Bool
    -fplugin-opt=Imp:--alias=Data.Char:Char
    -fplugin-opt=Imp:--alias=Data.Either:Either
    -fplugin-opt=Imp:--alias=Data.Map:Map
    -fplugin-opt=Imp:--alias=Data.Void:Void
    -fplugin-opt=Imp:--alias=System.IO:IO
```

So I felt like I was stacking the deck against the non-cached version. But best case it turned out to be about 1% faster:

```
without cache
cabal build  22.73s user 13.07s system 85% cpu 42.119 total
cabal build  22.64s user 13.05s system 84% cpu 42.063 total
cabal build  22.64s user 12.99s system 84% cpu 41.997 total
~22.67s

with cache
cabal build  22.70s user 13.14s system 84% cpu 42.409 total
cabal build  22.60s user 12.99s system 85% cpu 41.819 total
cabal build  22.46s user 12.86s system 85% cpu 41.234 total
~22.59s
```