@tfausak 2024-04-17
I'm going to close this for now. It might be worthwhile to resurrect if the performance benefits can be proved.